refactor(generators): refactor to make more testable
